+++ CHANGELOG.rst
@@ -18,6 +18,7 @@ Bug fixes:
 
 Other changes:
 
+- ``make_object`` is only called when input data are completely valid (:issue:`243`). Thanks :user:`kissgyorgy` for reporting.
 - Change default error messages for ``URL`` and ``Email`` validators so that they don't include user input (:issue:`255`).
 - ``Email`` validator permits email addresses with non-ASCII characters, as per RFC 6530 (:issue:`221`). Thanks :user:`lextoumbourou` for reporting and :user:`mwstobo` for sending the patch.